Update gptj-sagemaker.md
Fixing the max_length documentation. Taking the string length doesn't reflect the true stop length of max tokens. We should be taking the length of the tokens. Taking length of string would be 296 chars vs tokens 83 chars. When I was deploying my model, this was causing me timeout issues on max token generations.