The previous code does not work with the latest version (#762)
The previous code doesn't work with the latest version of transformers. It generated random text when we provide the model with deocoder input_id's and attention mask. Seems like the latest version implements shift right on input_ids automatically. Please refer to the link for clarification : https://github.com/huggingface/transformers/issues/18958 

The current version works for me with codebert as input and gpt2 as output. Thank you.